+++ packages/editor/src/app/pages/workspace/components/overlay/asset-browser/asset-browser.tsx
@@ -182,6 +182,26 @@ export const AssetDrawerElement = (
   }, [isCopying]);
 
   useEffect(() => {
+    const handleControls = (ev: KeyboardEvent) => {
+      switch (ev.code) {
+        case 'Escape':
+          onClose();
+          break;
+      }
+    };
+
+    if (isOpen) {
+      window.addEventListener('keydown', handleControls);
+    } else {
+      window.removeEventListener('keydown', handleControls);
+    }
+
+    return () => {
+      window.removeEventListener('keydown', handleControls);
+    };
+  });
+
+  useEffect(() => {
     searchAssetList();
   }, [filterInput, sortField, sortOrder]);