+++ packages/editor/src/app/pages/workspace/components/overlay/publish/course-settings.tsx
@@ -158,10 +158,11 @@ export const CourseSettings = ({
           className={`form-control form-control-sm${
             errors.name ? ' is-invalid' : ''
           }`}
+          value={name}
           placeholder={placeholderName}
           onChange={handleUpdateName}
           onKeyDown={handleInputName}
-          onBlur={handleFormUpdate}
+          onFocus={handleFormUpdate}
         />
         {errors.name && <div className="invalid-feedback">{errors.name}</div>}
       </div>
@@ -177,7 +178,7 @@ export const CourseSettings = ({
           value={description}
           onChange={handleUpdateDescription}
           onKeyDown={handleInputDescription}
-          onBlur={handleFormUpdate}
+          onFocus={handleFormUpdate}
         ></textarea>
       </div>
       <div className="mb-2">
@@ -193,7 +194,7 @@ export const CourseSettings = ({
           value={authors}
           onChange={handleUpdateAuthors}
           onKeyDown={handleInputAuthors}
-          onBlur={handleFormUpdate}
+          onFocus={handleFormUpdate}
         />
       </div>
       <div className="mb-2">
@@ -209,7 +210,7 @@ export const CourseSettings = ({
           value={organization}
           onChange={handleUpdateOrganization}
           onKeyDown={handleInputOrganization}
-          onBlur={handleFormUpdate}
+          onFocus={handleFormUpdate}
         />
       </div>
     </Accordion>