+++ packages/player/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.86 (2023-01-19)
+
+**Note:** Version bump only for package @scrowl/player
+
+
+
+
+
+## 0.0.85 (2023-01-19)
+
+**Note:** Version bump only for package @scrowl/player
+
+
+
+
+
+## 0.0.84 (2023-01-19)
+
+**Note:** Version bump only for package @scrowl/player
+
+
+
+
+
 ## 0.0.83 (2023-01-11)
 
 **Note:** Version bump only for package @scrowl/player